backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

opencv-python-headless==4.11.0.86
numpy>=1.26.0,<2.0.0
Pillow==11.1.0

paddlepaddle==2.6.2
paddleocr==2.9.1

insightface==0.7.3
onnxruntime-gpu==1.20.1

openai==1.58.1
python-Levenshtein==0.26.1

aiofiles==24.1.0
